backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
music21>=9.1.0
basic-pitch>=0.3.0
oemer>=0.1.6
numpy>=1.24.0
scipy>=1.11.0
librosa>=0.10.0
soundfile>=0.12.1
pydantic>=2.0.0
aiofiles>=23.2.1
Pillow>=10.0.0
